refactor splitting up resources.js into more manageable modules
