+++ util/constants.js
@@ -1,0 +1,1 @@
+export const allowedUserRoles = ['super_user', 'read_only_user'];
